requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
httpx>=0.25.0
python-dotenv>=1.0.0
supabase>=2.0.3
PyPDF2>=3.0.1
beautifulsoup4>=4.12.2
google-generativeai>=0.3.1
python-multipart>=0.0.6
pydub>=0.25.1
uuid>=1.30
openai>=1.3.5  # Optional for OpenAI TTS
email-validator>=2.0.0  # Required for validating email fields